requirements.txt
Flask==2.2
Werkzeug==2.2
opencv-python-headless
tensorflow-cpu==2.10.0
numpy==1.23.0
pillow==9.2.0
gunicorn==20.1.0
Flask-SQLAlchemy==2.5.1
pandas==1.5.0
matplotlib==3.6.1
scikit-learn==1.1.2
tqdm==4.64.1